\name{Hind.QuartetEffect}
\alias{Hind.QuartetEffect}
\alias{Hind.QuartetEffect}
\non_function{}
\title{ Hind.QuartetEffect}
\usage{
 
}
\description{
 It contains computed quartet effect for Hind array. 
 Those were computed on some normal samples of the hapmap dataset.
 
}
\format{

}
\source{}
\references{
}

\keyword{datasets}

