\name{Nsp.QuartetEffect}
\alias{Nsp.QuartetEffect}
\alias{Nsp.QuartetEffect}
\non_function{}
\title{ Nsp.QuartetEffect}
\usage{
 
}
\description{
 It contains computed quartet effect for Nsp array. 
 Those were computed on some normal samples of the hapmap dataset.
 
}
\format{

}
\source{}
\references{
}

\keyword{datasets}
