\name{Sty.QuartetEffect}
\alias{Sty.QuartetEffect}
\alias{Sty.QuartetEffect}
\non_function{}
\title{ Sty.QuartetEffect}
\usage{
 
}
\description{
 It contains computed quartet effect for Sty array. 
 Those were computed on some normal samples of the hapmap dataset.
 
}
\format{

}
\source{}
\references{
}

\keyword{datasets}
