\name{Xba.QuartetEffect}
\alias{Xba.QuartetEffect}
\alias{Xba.QuartetEffect}
\non_function{}
\title{ Xba.QuartetEffect}
\usage{
 
}
\description{
 It contains computed quartet effect for Xba array. 
 Those were computed on some normal samples of the hapmap dataset.
 
}
\format{

}
\source{}
\references{
}

\keyword{datasets}
